rename module neighbor.py to knn_graph.py
