+++ cleanlab/internal/neighbor/__init__.py
@@ -1,1 +1,1 @@
-from .neighbor import features_to_knn
+from .knn_graph import features_to_knn
